Fixes for standard raptor gcc -W args
raptor calls gcc with (at least) the following warnings:
gcc  -Wall -Wextra -Wshadow -Wpointer-arith -Wstrict-prototypes
-Wmissing-prototypes -Wmissing-declarations -Wnested-externs
-Wredundant-decls -Wswitch-enum -Wsign-compare
-Werror-implicit-function-declaration -Wwrite-strings -Wpacked
-Wmissing-format-attribute -Wpointer-arith -Wcast-align -Winit-self
-Wdeclaration-after-statement -Wold-style-definition
-Wno-missing-field-initializers -Wno-unused-parameter
-Wformat-security

Some such as 'attribute' => 'attr' are unfortunate but
are due to shadowed function prototype declarations in
some headers that gcc reports as noisy warnings.

+++ c/subject.c
@@ -39,7 +39,7 @@ char* rdfa_create_bnode(rdfacontext* context)
    char buffer[64];
 
    // print and increment the bnode count
-   sprintf(buffer, "_:bnode%i", context->bnode_count++);
+   sprintf(buffer, "_:bnode%i", (int)context->bnode_count++);
    rval = rdfa_replace_string(rval, buffer);
 
    return rval;
